back-end/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.7.2
blinker==1.9.0
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.116.1
Flask==3.1.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
googletrans==3.0.0
greenlet==3.2.3
gyp==0.1
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.20.2
httpx==0.13.3
hyperframe==5.2.0
idna==3.10
importlib-metadata==4.6.4
iniconfig==2.0.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==3.0.2
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.2
packaging==23.2
pillow==11.3.0
pluggy==1.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2==2.9.9
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.10.1
pyparsing==2.4.7
pyperclip==1.9.0
pytest==7.3.1
pytest-mock==3.10.0
pytest-print==0.3.1
python-apt==2.4.0+ubuntu4.1
python-dotenv==1.1.1
python-jose==3.5.0
pytz==2023.3.post1
PyYAML==5.4.1
redgreenunittest==0.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9.1
SecretStorage==3.3.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
sqlparse==0.4.4
starlette==0.47.2
systemd-python==234
tomli==2.0.1
typing-inspection==0.4.1
typing_extensions==4.14.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.1.0
uuid==1.30
uvicorn==0.35.0
wadllib==1.3.6
watchdog==6.0.0
Werkzeug==3.1.3
wsproto==1.0.0
zipp==1.0.0
